Struct gfx_core::command::AccessGuard

pub struct AccessGuard<'a, R: Resources> { /* fields omitted */ }

Methods

impl<'a, R: Resources> AccessGuard<'a, R>

fn access_mapped_reads(&mut self) -> AccessGuardBuffers<R>

Returns the mapped buffers that The GPU will read from, with exclusive acces to their mapping

fn access_mapped_writes(&mut self) -> AccessGuardBuffers<R>

Returns the mapped buffers that The GPU will write to, with exclusive acces to their mapping

fn access_mapped(&mut self) -> AccessGuardBuffersChain<R>