Module gfx_core::dummy

Dummy backend implementation to test the code for compile errors outside of the graphics development environment.

Structs

DummyCommandBuffer
Dummy command buffer, which ignores all the calls.

DummyDevice
Dummy device which does minimal work, just to allow testing gfx-rs apps for compilation.

DummyFence
Dummy fence that does nothing.

DummyMapping
Dummy mapping which will crash on use.

Enums

DummyResources
Dummy resources phantom type