Enum gfx_core::format::SurfaceType

#[repr(u8)]

pub enum SurfaceType {
    R4_G4,
    R4_G4_B4_A4,
    R5_G5_B5_A1,
    R5_G6_B5,
    R8,
    R8_G8,
    R8_G8_B8_A8,
    R10_G10_B10_A2,
    R11_G11_B10,
    R16,
    R16_G16,
    R16_G16_B16,
    R16_G16_B16_A16,
    R32,
    R32_G32,
    R32_G32_B32,
    R32_G32_B32_A32,
    B8_G8_R8_A8,
    D16,
    D24,
    D24_S8,
    D32,
}

Type of the allocated texture surface. It is supposed to only carry information about the number of bits per each channel. The actual types are up to the views to decide and interpret. The actual components are up to the swizzle to define.

Variants
R4_G4R4_G4_B4_A4R5_G5_B5_A1R5_G6_B5R8R8_G8R8_G8_B8_A8R10_G10_B10_A2R11_G11_B10R16R16_G16R16_G16_B16R16_G16_B16_A16R32R32_G32R32_G32_B32R32_G32_B32_A32B8_G8_R8_A8D16D24D24_S8D32
Methods

impl SurfaceType

pub fn get_total_bits(&self) -> u8

Return the total number of bits for this format.

pub fn get_alpha_stencil_bits(&self) -> u8

Return the number of bits allocated for alpha and stencil.