Module gfx_core::memory

Types to describe the properties of memory allocated for gfx resources.

Structs

Access
Flags providing information about the type of memory access to a resource.

Bind
Flags providing information about the usage of a resource.

Enums

Usage
How this memory will be used regarding GPU-CPU data flow.

Traits

Pod
A trait for plain-old-data types.

Functions

cast_slice
Cast a slice from one POD type to another.