Module gfx_core::pso

Pipeline State Objects

Structs

ColorInfo
Color output configuration of the PSO.

ConstantBufferSet
A complete set of constant buffers to be used for the constants binding in PSO.

CreationError
Error types happening upon PSO creation on the device side.

DepthStencilInfo
Depth and stencil state of the PSO.

Descriptor
All the information surrounding a shader program that is required for PSO creation, including the formats of vertex buffers and pixel targets;

Element
A struct element descriptor.

PixelTargetSet
A complete set of render targets to be used for pixel export in PSO.

ResourceViewSet
A complete set of shader resource views to be used in PSO.

SamplerSet
A complete set of samplers to be used for PSO.

UnorderedViewSet
A complete set of unordered access views to be used in PSO.

VertexBufferSet
A complete set of vertex buffers to be used for vertex import in PSO.

Type Definitions

AttributeDesc
PSO vertex attribute descriptor

BufferOffset
An offset inside a vertex buffer, in bytes.

ColorTargetDesc
PSO color target descriptor

DepthStencilDesc
PSO depth-stencil target descriptor

ElemOffset
Offset of an attribute from the start of the buffer, in bytes

ElemStride
Offset between attribute values, in bytes

InstanceRate
The number of instances between each subsequent attribute value