Struct gfx_core::pso::AccessInfo

pub struct AccessInfo<R: Resources> { /* fields omitted */ }

Informations about what is accessed by the pipeline

Methods

impl<R: Resources> AccessInfo<R>

fn new() -> Self

Creates empty access informations

fn clear(&mut self)

Clear access informations

fn buffer_read(&mut self, buffer: &RawBuffer<R>)

Register a buffer read access

fn buffer_write(&mut self, buffer: &RawBuffer<R>)

Register a buffer write access

fn mapped_reads(&self) -> &[RawBuffer<R>]

Returns a slice of mappings associated to buffers that The GPU will read from

fn mapped_writes(&self) -> &[RawBuffer<R>]

Returns a slice of mappings associated to buffers that The GPU will write to