Struct gfx_core::pso::Descriptor

pub struct Descriptor {
    pub primitive: Primitive,
    pub rasterizer: Rasterizer,
    pub scissor: bool,
    pub vertex_buffers: [Option<VertexBufferDesc>; 4],
    pub attributes: [Option<AttributeDesc>; 16],
    pub constant_buffers: [Option<ConstantBufferDesc>; 14],
    pub resource_views: [Option<ResourceViewDesc>; 32],
    pub unordered_views: [Option<UnorderedViewDesc>; 4],
    pub samplers: [Option<SamplerDesc>; 16],
    pub color_targets: [Option<ColorTargetDesc>; 4],
    pub depth_stencil: Option<DepthStencilDesc>,
}

All the information surrounding a shader program that is required for PSO creation, including the formats of vertex buffers and pixel targets;

Fields
primitive: Primitive

Type of the primitive

rasterizer: Rasterizer

Rasterizer setup

scissor: bool

Enable scissor test

vertex_buffers: [Option<VertexBufferDesc>; 4]

Vertex buffers

attributes: [Option<AttributeDesc>; 16]

Vertex attributes

constant_buffers: [Option<ConstantBufferDesc>; 14]

Constant buffers

resource_views: [Option<ResourceViewDesc>; 32]

Shader resource views

unordered_views: [Option<UnorderedViewDesc>; 4]

Unordered access views

samplers: [Option<SamplerDesc>; 16]

Samplers

color_targets: [Option<ColorTargetDesc>; 4]

Render target views (RTV)

depth_stencil: Option<DepthStencilDesc>

Depth stencil view (DSV)

Methods

impl Descriptor

pub fn new(primitive: Primitive, rast: Rasterizer) -> Descriptor

Create a new empty PSO descriptor.