Enum gfx_core::tex::Kind

pub enum Kind {
    D1(Size),
    D1Array(Size, ArraySize),
    D2(Size, Size, AaMode),
    D2Array(Size, Size, ArraySize, AaMode),
    D3(Size, Size, Size),
    Cube(Size, Size),
    CubeArray(Size, Size, ArraySize),
}

Specifies the kind of a texture storage to be allocated.

Variants

D1

A single row of texels.

D1Array

An array of rows of texels. Equivalent to Texture2D except that texels in a different row are not sampled.

D2

A traditional 2D texture, with rows arranged contiguously.

D2Array

An array of 2D textures. Equivalent to Texture3D except that texels in a different depth level are not sampled.

D3

A volume texture, with each 2D layer arranged contiguously.

Cube

A set of 6 2D textures, one for each face of a cube.

CubeArray

An array of Cube textures.

Methods

impl Kind

fn get_dimensions(&self) -> Dimensions

Get texture dimensions, with 0 values where not applicable.

fn get_level_dimensions(&self, level: Level) -> Dimensions

Get the dimensionality of a particular mipmap level.