Enum gfx_core::texture::CreationError

pub enum CreationError {
    Format(SurfaceType, Option<ChannelType>),
    Kind,
    Samples(AaMode),
    Size(Size),
    Data(usize),
    Usage(Usage),
    Mipmap,
    Level(Level),
}

Pure texture object creation error.

Variants
Format(SurfaceType, Option<ChannelType>)

Failed to map a given format to the device.

Kind

The kind doesn't support a particular operation.

Samples(AaMode)

Failed to map a given multisampled kind to the device.

Size(Size)

Unsupported size in one of the dimensions.

Data(usize)

The given data has a different size than the target texture slice.

Usage(Usage)

The mentioned usage mode is not supported

Mipmap

The requested mipmap creation parameter is unsupported.

Level(Level)

The requested mipmap level count does not match the provided data.