Trait gfx_core::memory::Pod

pub unsafe trait Pod { }

A trait for plain-old-data types.

A POD type does not have invalid bit patterns and can be safely created from arbitrary bit pattern. The Pod trait is implemented for standard integer and floating point numbers as well as common arrays of them (for example [f32; 2]).