Module gfx_core::pso

Raw Pipeline State Objects

This module contains items used to create and manage a raw pipeline state object. Most users will want to use the typed and safe PipelineState. See the pso module inside the gfx crate.

Structs

ColorInfo
Color output configuration of the PSO.

ConstantBufferParam
A constant buffer run-time parameter for PSO.

CreationError
Error types happening upon PSO creation on the device side.

DepthStencilInfo
Depth and stencil state of the PSO.

Descriptor
All the information surrounding a shader program that is required for PSO creation, including the formats of vertex buffers and pixel targets;

Element
A struct element descriptor.

PixelTargetSet
A complete set of render targets to be used for pixel export in PSO.

ResourceViewParam
A shader resource view (SRV) run-time parameter for PSO.

SamplerParam
A sampler run-time parameter for PSO.

UnorderedViewParam
An unordered access view (UAV) run-time parameter for PSO.

VertexBufferDesc
Vertex buffer descriptor

VertexBufferSet
A complete set of vertex buffers to be used for vertex import in PSO.

Constants

MAX_VERTEX_BUFFERS
Maximum number of vertex buffers used in a PSO definition.

Type Definitions

AttributeDesc
PSO vertex attribute descriptor

BufferIndex
Index of a vertex buffer.

BufferOffset
An offset inside a vertex buffer, in bytes.

ColorTargetDesc
PSO color target descriptor

ConstantBufferDesc
PSO constant buffer descriptor

DepthStencilDesc
PSO depth-stencil target descriptor

ElemOffset
Offset of an attribute from the start of the buffer, in bytes

ElemStride
Offset between attribute values, in bytes

InstanceRate
The number of instances between each subsequent attribute value

ResourceViewDesc
PSO shader resource view descriptor

SamplerDesc
PSO sampler descriptor

UnorderedViewDesc
PSO unordered access view descriptor