Struct gfx_core::pso::PixelTargetSet

pub struct PixelTargetSet<R: Resources> {
    pub colors: [Option<R::RenderTargetView>; 4],
    pub depth: Option<R::DepthStencilView>,
    pub stencil: Option<R::DepthStencilView>,
    pub dimensions: Option<Dimensions>,
}

A complete set of render targets to be used for pixel export in PSO.

Fields
colors: [Option<R::RenderTargetView>; 4]

Array of color target views

depth: Option<R::DepthStencilView>

Depth target view

stencil: Option<R::DepthStencilView>

Stencil target view

dimensions: Option<Dimensions>

Rendering dimensions

Methods

impl<R: Resources> PixelTargetSet<R>

fn new() -> PixelTargetSet<R>

Create an empty set

fn add_color(&mut self,
slot: ColorSlot,
view: &R::RenderTargetView,
dim: Dimensions)

Add a color view to the specified slot

fn add_depth_stencil(&mut self,
view: &R::DepthStencilView,
has_depth: bool,
has_stencil: bool,
dim: Dimensions)

Add a depth or stencil view to the specified slot

fn get_view(&self) -> (u16, u16, u16)

Get the rendering view (returns values > 0)